ref: Reorganize Day 3 files
