+++ lib/puzzles/2023/day03/input.txt
